+++ mopidy/frontends/mpd/protocol/empty.py
@@ -1,9 +1,0 @@
-from __future__ import unicode_literals
-
-from mopidy.frontends.mpd.protocol import handle_request
-
-
-@handle_request(r'[\ ]*$')
-def empty(context):
-    """The original MPD server returns ``OK`` on an empty request."""
-    pass